requirements.txt
asgiref==3.5.2
Django==4.1.3
django-environ==0.11.2
pytest==7.0.0
pytest-django==4.7.0
python-decouple==3.8
pytz==2021.1
djangorestframework==3.11.2
mysqlclient==2.1.1
sqlparse==0.4.3
tzdata==2022.7
